+++ CHANGELOG.md
@@ -1,5 +1,9 @@
-## 0.0.8
-* Option -i|--install is optionnal, deps are checked during the execution 
+## 0.0.8, release 2020-05-06
+* Always make a copy and replace original file, do not ask, the program is dependent of his config file for now.
+* Correct install where Dir.chdir don't back to the previous location.
+* --install also place the deceitmac.service if your system use systemd.
+* Simplify lib/spior/copy check\_hash.
+* The use of the option -i|--install is optionnal, deps are checked during the execution.
 
 ## 0.0.7, release 2020-05-05
 * Execute the option -t|--tor verify if config files are present.